breaking change to dataframe schema API: use dict
dataframe schema is now specified with a dictionary
where keys are column names and values are Column objects.
This new API can easily extend to multi-column schema checking
as well as multiindex columns, and even multi-index, multi-column
schema checking.

Renamed the `Validator` class to `Check` for brevity and clarity,
and created a variable name for each PandasDtype at the top
of the pandera namespace