add `groupby` and `groups` arg to `Check`
fixes #16.

this diff enables more complex `Column` `Check`s that incorporate
data from other columns in a principled way.

This is done by providing a `groupby` argument, which enables
the user to make assertions about subsets of the column based
on different groups. This is the first step towards #17.

This diff also changes the default of `element_wise=False`