add PandasDtype.Str to differentiate native pandas str type
fixes #312

this diff adds a `Str` enum to the `PandasDtype` enum class so that
there's a distinction between the str dtype that pandas represents
as an object array and the native string dtype that became available
after pandas v1. For <1 pandas installations, the PandasDtype.String
and "string" alias will fall back on the PandasDtype.Str dtype.