bugfix: coerce nullable str column handles all na
this diff fixes an issue reported in #365 that fails to coerce
nullable string column in the case that all values are null.
The issue surfaces when reading a csv file when a column has
all null values, since these columns will be of float dtype. We
can reproduce this behavior by creating a null column in a
dataframe of float type.

Add tests to make sure this case is handled, along with the
nullable string and integer types in pandas>1.0.0.