pin pandas-stubs version, bump mypy
as of this commit, the latest version of `pandas-stubs == 1.4.3.220915`
causes a breaking change when using `pandera.typing.DataFrame[Schema]`
to initialize validated dataframes. The issue is tracked in the
pandas-stubs repo: https://github.com/pandas-dev/pandas-stubs/issues/197

In the mean time, we pin the pandas-stubs to the previous available
version, where this doesn't show up.

Signed-off-by: Niels Bantilan <niels.bantilan@gmail.com>