De/serializate data frame schema to/from JSON
* Make the de/serialization from/to JSON-able mappings public so that
  we allow for other JSON-like de/serializations in the future.
* Add functions for de/serializing the data frame schema to/from JSON
  files, streams and strings.

Note to future developers: it would make sense to refactor the tests so
that the de/serialization of JSON-able mappings is tested in isolation
to target format (at the moment: JSON and YAML). Currently, the tests
are coupled with the YAML format. The authors of this patch did not have
time cycles to refactor the tests, but there is no reason not to do it
(except for the lack of time).

Fixes #918.

Co-authored-by: Marko Ristin-Kaufmann <marko@ristin.ch>