make df strategy less complex
fixes #988

A change in #658 introduced a step to handle str/object dtypes due
to issues handling np.str_. Will need to look into whether that's
still necessary in another PR, but this one compresses a bunch
of `strategy.map` calls to a single one.

This addresses an issue where the strategy would be way too long
for schemas with many str/object columns.

Signed-off-by: Niels Bantilan <niels.bantilan@gmail.com>