feature: Copy columns in DataFrameSchema __init__
Fixes https://github.com/unionai-oss/pandera/issues/511

Also restructures so that the conversion is a pure function,
which is easier to reason about
and is less prone to breaking in the future
from unintentional side effects from
self.

Also move all verifications to top of __init__, so all the
assignments to self happen in one block.

Also contains a small typing fix from previous commit.

Signed-off-by: Nick Crews <nicholas.b.crews@gmail.com>